% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkDateLimits}
\alias{checkDateLimits}
\title{Check if input falls within Date limits for currently supported ERDDAP data}
\usage{
checkDateLimits(type, firstDay)
}
\arguments{
\item{type}{Type of environmental layer. Currently
takes 3 possible values: \code{'productivity'}, \code{'chlorophyll'} and 
\code{'sst'}. See Details below.}

\item{firstDay}{An object of class \code{\link[base]{Date}} of format YYYY-MM-01.}
}
\value{
An object of class \code{\link[base]{Date}} of format 'YYYY-MM-DD'
currently supported ERDDAP data
}
\description{
Check if input falls within Date limits
}
\details{
The function uses internal \code{\link[noaaErddap]{dataTypeLimits}} to
evaluate whether the provided month and year in \code{\link[noaaErddap]{erddapDownload}}
follow within the supported dates. If YYYY-MM is identical to either beginning  
or end of supported range, than the function modifies \code{firstDay} to have the same
day as the starting (or ending) supported date.
}
\seealso{
\code{\link[noaaErddap]{dataTypeLimits}}, \code{\link[noaaErddap]{erddapDownload}}.
}
\author{
Diego Barneche.
}
