% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dataTypeLimits}
\alias{dataTypeLimits}
\title{Date limits for currently supported ERDDAP data}
\usage{
dataTypeLimits()
}
\value{
A \code{\link[base]{data.frame}} showing the Date limits 
currently supported ERDDAP data
}
\description{
Date limits for currently supported ERDDAP data
}
\author{
Diego Barneche.
}
