% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{erddapGet}
\alias{erddapGet}
\title{Create a download link for SST data}
\usage{
erddapGet(type, filePath, firstDay, overwrite, ...)
}
\arguments{
\item{type}{Type of environmental layer. Currently
takes 3 possible values: \code{'productivity'}, \code{'chlorophyll'} and 
\code{'sst'}.}

\item{filePath}{The file path within the cached as created by \code{\link[noaaErddap]{noaaErddapCacheDir}}}

\item{firstDay}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD.}

\item{overwrite}{Should the layer of interest be overwritten?}

\item{...}{further arguments to \code{\link[httr]{GET}}.}
}
\value{
A character vector containing the path where files have been saved to.
}
\description{
Downloads data to package cache folder
}
\seealso{
\code{\link[noaaErddap]{erddapDownload}}, \code{\link[noaaErddap]{noaaErddapCacheDir}}.
}
\author{
Diego Barneche.
}
