% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{erddapLocal}
\alias{erddapLocal}
\title{Creates cache folder to store ERDDAP datasets}
\usage{
erddapLocal(path, firstDay)
}
\arguments{
\item{path}{A folder path created via \link[noaaErddap]{noaaErddapCacheDir}.}

\item{firstDay}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD.}
}
\value{
A file path.
}
\description{
Cache folder for ERDDAP datasets
}
\details{
Uses path created via \link[noaaErddap]{noaaErddapCacheDir} 
to create a full path in cache folder to store ERDDAP datasets.
}
\seealso{
\code{\link[noaaErddap]{noaaErddapCacheDir}}.
}
\author{
Diego Barneche.
}
