% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extractDataWithRaster}
\alias{extractDataWithRaster}
\title{Transform NetCDF dates into human dates}
\usage{
extractDataWithRaster(slices, filePath, coordinates, ...)
}
\arguments{
\item{slices}{A data.frame frame with column \code{band} indicating
which layer is to be extracted.}

\item{filePath}{A path to a NetCDF file.}

\item{coordinates}{A data.frame containing columns \code{Longitude} and \code{Latitude}
indicating the coordinates of interest.}

\item{...}{Additional arguments to \link[raster]{extract}.}
}
\value{
A matrix (longitude, latitude, time) containing variable of interest.
}
\description{
Transform NetCDF dates into human dates
}
\author{
Diego Barneche.
}
