% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findLastDayOfTheMonth}
\alias{findLastDayOfTheMonth}
\title{Finds the last day of the month}
\usage{
findLastDayOfTheMonth(myDate)
}
\arguments{
\item{myDate}{An object of class \code{\link[base]{Date}}.}
}
\value{
A \code{\link[base]{character}} vector of format YYYY-MM-DD.
}
\description{
Last day of the month
}
\seealso{
\code{\link[noaaErddap]{getFirstDayOfTheMonth}}.
}
\author{
Diego Barneche, code adapted from \href{https://www.r-bloggers.com/find-the-last-day-of-the-month/}{R-Bloggers}.
}
