% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getFirstDayOfTheMonth}
\alias{getFirstDayOfTheMonth}
\title{Concatenates month and year to generate the first day of the month}
\usage{
getFirstDayOfTheMonth(month, year)
}
\arguments{
\item{month}{A \code{\link[base]{numeric}} (integer) or \code{\link[base]{character}} string (MM) informing the month (from 1 to 12).}

\item{year}{A \code{\link[base]{numeric}} or \code{\link[base]{character}} string (YYYY) informing the year.}
}
\value{
An object of class \code{\link[base]{Date}} of format YYYY-MM-01.
}
\description{
First day of the month
}
\seealso{
\code{\link[noaaErddap]{findLastDayOfTheMonth}}.
}
\author{
Diego Barneche.
}
