% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{noaaErddapCacheDir}
\alias{noaaErddapCacheDir}
\title{Creates cache folder to store ERDDAP datasets}
\usage{
noaaErddapCacheDir()
}
\description{
Cache folder for ERDDAP datasets
}
\details{
Uses \link[rappdirs]{user_cache_dir} to create
a cache folder to store ERDDAP datasets
}
\seealso{
\code{\link[noaaErddap]{erddapLocal}}.
}
\author{
Diego Barneche.
}
