% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{noaaErddapFiles}
\alias{noaaErddapFiles}
\title{Provides a vector with all currently-stored ERDDAP datasets}
\usage{
noaaErddapFiles(type, ...)
}
\arguments{
\item{type}{Type of environmental layer. Currently
takes 3 possible values: \code{'productivity'}, \code{'chlorophyll'} and 
\code{'sst'}.}

\item{...}{further arguments to \code{\link[base]{dir}}.}
}
\description{
Vector of ERDDAP datasets
}
\details{
Searches cache folder to find currently stored ERDDAP datasets
}
\author{
Diego Barneche.
}
