% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{openAndMatchNcdfData}
\alias{openAndMatchNcdfData}
\title{Extract NetCDF data}
\usage{
openAndMatchNcdfData(filePath, method = c("ncdf4", "raster"), dayFilter,
  coordinates, ...)
}
\arguments{
\item{filePath}{A path to a NetCDF file.}

\item{method}{Method (package) of extraction. Currently supports
either \code{ncdf4} or \code{raster}. See Details below.}

\item{dayFilter}{Optional. A vector containing Date objects with
dates of interest.}

\item{coordinates}{Optional, only used if method is \code{raster}.
A data.frame containing columns \code{Longitude} and \code{Latitude}
indicating the coordinates of interest.}

\item{...}{Additional arguments to \link[raster]{extract} (used internally).}
}
\value{
An array (longitude, latitude, time) containing variable of interest.
}
\description{
Extract NetCDF data
}
\details{
Method \code{ncdf4} uses the package to extract all layers using all coordinates.
Method \code{raster} allows the user to specify arguments to the raster function
\link[raster]{extract} together with a table of coordinates of interest.
}
\examples{
library(noaaErddap)
library(plyr)

nppFiles  <-  noaaErddapFiles('productivity', full.name = TRUE)

# extract NPP values for a given subset of coordinates (median value within a buffer of 20 km) across all files and take the mean
nppValues  <-  abind(lapply(nppFiles, openAndMatchNcdfData, method = 'raster', coordinates = data.frame(Longitude = c(330, 335, 340), #' Latitude = c(-27, -19, 0)), buffer = 2e4, fun = median, na.rm = TRUE), along = 3)
apply(nppValues, c(1, 2), mean, na.rm = TRUE)

# extract the average NPP values for the globe in 1998
nppValues1998  <-  abind(lapply(nppFiles[grep('-1998.nc', nppFiles, fixed = TRUE)], openAndMatchNcdfData, method = 'ncdf4'), along = 3)
meanNPP1998    <-  apply(nppValues1998, c(1, 2), mean, na.rm = TRUE)
}
\seealso{
\code{\link[noaaErddap]{noaaErddapFiles}}.
export
}
\author{
Diego Barneche.
}
