% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{productivityBaseLink}
\alias{productivityBaseLink}
\title{Create a download link for productivity data}
\usage{
productivityBaseLink(firstDay)
}
\arguments{
\item{firstDay}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD.}
}
\value{
An object of class \code{\link[base]{character}}.
}
\description{
Base download link for productivity data
}
\seealso{
\code{\link[noaaErddap]{erddapGet}}.
}
\author{
Diego Barneche.
}
