% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sstBaseLink}
\alias{sstBaseLink}
\title{Create a download link for SST data}
\usage{
sstBaseLink(firstDay)
}
\arguments{
\item{firstDay}{An object of class \code{\link[base]{Date}} of format YYYY-MM-DD.}
}
\value{
An object of class \code{\link[base]{character}}.
}
\description{
Base download link for SST data
}
\seealso{
\code{\link[noaaErddap]{erddapGet}}.
}
\author{
Diego Barneche.
}
