% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transformDateForLayer}
\alias{transformDateForLayer}
\title{Transform NetCDF dates into human dates}
\usage{
transformDateForLayer(envNc, type)
}
\arguments{
\item{envNc}{A file of class \code{\link[ncdf4]ncdf4}.}

\item{type}{Type of environmental layer. Currently
takes 3 possible values: \code{'productivity'}, \code{'chlorophyll'} and 
\code{'sst'}.}
}
\value{
A vector of class \code{\link[base]{Date}} of format YYYY-MM-DD.
}
\description{
Transform NetCDF dates into human dates
}
\author{
Diego Barneche.
}
