% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{typeFromPath}
\alias{typeFromPath}
\title{Extracts type of ERDDAP datasets based on
file path.}
\usage{
typeFromPath(filePath)
}
\arguments{
\item{filePath}{A path to a NetCDF file.}
}
\value{
A character vector containing the type of variable.
}
\description{
Type of ERDDAP datasets
}
\author{
Diego Barneche.
}
