% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{erddapDownload}
\alias{erddapDownload}
\title{Download SST, NPP, and Chlorophyll Data from NOAA (ERDDAP)}
\usage{
erddapDownload(year, month, type, overwrite = FALSE, ...)
}
\arguments{
\item{year}{A numeric string (XXXX) informing the year.}

\item{month}{A numeric string informing the month (from 1 to 12).}

\item{type}{Type of environmental layer to be downloaded. Currently
takes 3 possible values: \code{'productivity'}, \code{'productivity'} and 
\code{'productivity'}. See \code{details} below.}

\item{overwrite}{Should the layer of interest be overwritten?}

\item{...}{further arguments to \code{httr::GET}}
}
\value{
A character vector containing the path where files have been saved to.
}
\description{
Download NOAA ERDDAP Data.
}
\details{
Net Primary Productivity (NPP) data
from http://coastwatch.pfeg.noaa.gov/erddap/griddap/erdPPbfp18day.html
NOAA ERDDAP > griddap > Data Access Form 
metadata in http://coastwatch.pfeg.noaa.gov/erddap/info/erdPPbfp18day/index.html 
Primary Productivity, SeaWiFS and Pathfinder, Global, 1997-2010, EXPERIMENTAL (8 Day Composite)
From 1997 - Nov - 29 to 2007 - Nov - 29
NPP is an 8-day composite product, so every consecutive eight days should produce an independent
time step. Note that this time series is not perfectly even. 
Data input should be in the format of year and month.
Chlorophyll-a (CHL) data
from http://coastwatch.pfeg.noaa.gov/erddap/griddap/erdSW1chla8day.html
NOAA ERDDAP > griddap > Data Access Form 
metadata in http://coastwatch.pfeg.noaa.gov/erddap/info/erdSW1chla8day/index.html
Chlorophyll-a, Orbview-2 SeaWiFS, R2014.0, 0.1°, Global, 1997-2010 (8 Day Composite) 
From 1997 - Nov - 29 to 2007 - Nov - 29
Chlorophyll-a is an 8-day composite product, so every consecutive eight days should produce an independent
time step. Note that this time series is not perfectly even. 
Data input should be in the format of year and month.
Sea Surface Temperature (SST) data
from http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.highres.html
list of individual files available from:
http://www.esrl.noaa.gov/psd/cgi-bin/db_search/DBListFiles.pl?did=132&tid=55440&vid=2423
Brief Description: NOAA High-resolution Blended Analysis of Daily SST and Ice. 
Data is from Sep 1981 and is on a 1/4 deg global grid.
Temporal Coverage: Daily values from 1981/09 to present
Spatial Coverage: 0.25 degree latitude x 0.25 degree longitude global grid (1440x720).
89.875S - 89.875N,0.125E to 359.875E.
Main ref: Reynolds, Richard W., Thomas M. Smith, Chunying Liu, Dudley B. Chelton, Kenneth S. Casey, Michael G. Schlax, 2007: Daily H'igh-Resolution-Blended Analyses for Sea Surface Temperature. J. Climate, 20, 5473-5496. Reynolds, Richard W., Thomas M. Smith, Chunying Liu, Dudley B. Chelton, Kenneth S. Casey, Michael G. Schlax, 2007: Daily High-Resolution-Blended Analyses for Sea Surface Temperature. J. Climate, 20, 5473-5496.
}
\examples{
# download productivity data for Jan 2006
pathToDownloadedFile  <-  erddapDownload(year = 2006, month = 1, type = 'productivity')
# download productivity data for all months between 1998-2008
library(plyr)
dat  <-  expand.grid(year = 1998:2008, month = 1:12)
tabOfLinks  <-  ddply(dat, .(year, month), function(x, type)  {
    erddapDownload(x$year, x$month, type, overwrite = TRUE)
}, type = 'productivity')
# check outputs
pathToDownloadedFile; tabOfLinks
}
\author{
Diego Barneche.
}
